requirements/local.txt
# Local development dependencies go here
-r test.txt

django-debug-toolbar==1.2.2
